backend/requirements.txt
fastapi>=0.115.5
uvicorn[standard]>=0.32.1
sqlalchemy>=2.0.36
psycopg2-binary>=2.9.10
alembic>=1.14.0
pydantic[email]>=2.11.0
pydantic-settings>=2.7.0
email-validator>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.17
python-dotenv>=1.0.1
